backend/requirements.txt
requests
fastapi
uvicorn
httpx
supabase
pydantic
google-generativeai
google-genai
python-dotenv
